+++ resources/views/admin/article_list.blade.php
@@ -4,7 +4,11 @@
         $('#left-bar .nav>li:nth-child(4)>.menu').addClass('on');
     </script>
 
-
+    <h3 style="float: right;margin-top: 2%"><a href="@if( $pid == 0 )
+                /admin/article
+              @else
+                /article/{{ $pid }}/look
+               @endif">返回上级</a></h3>
     <div class="main-wrap fadeInRight">
         <p style="color: red"><?php
             if(!is_object($errors)){
@@ -30,7 +34,7 @@
                         <a class="act-btn" href="/admin/article/{{$sort->id}}/edit">编辑</a>|
                         <a class="act-btn" href="/article/{{$sort->id}}/look">查看内容</a>|
                         <a class="act-btn" href="/article/{{$sort->id}}/addson">添加内容</a>|
-                        <a class="act-btn" href="/article/{{$sort->id}}/delete">删除</a>
+                        <a class="act-btn"  onclick="if(confirm('确定删除此项？')){location.href='/article/{{$sort->id}}/delete'}"}>删除</a>
                     </td>
                 </tr>
             @endforeach
